+++ src/js/records/requests/asset-create.record.js
@@ -9,39 +9,42 @@ export class AssetCreateRequestRecord extends RequestRecord {
     super(record)
 
     this.assetName = _get(
-      this._record, 'details.assetCreate.details.name'
+      this._record, 'details.createAsset.details.name'
     )
     this.assetCode = _get(
-      this._record, 'details.assetCreate.code'
+      this._record, 'details.createAsset.code'
     )
     this.preissuedAssetSigner = _get(
-      this._record, 'details.assetCreate.preIssuedAssetSigner'
+      this._record, 'details.createAsset.preIssuedAssetSigner'
     )
     this.maxIssuanceAmount = _get(
-      this._record, 'details.assetCreate.maxIssuanceAmount'
+      this._record, 'details.createAsset.maxIssuanceAmount'
+    )
+    this.assetType = _get(
+      this._record, 'details.createAsset.type'
     )
     this.initialPreissuedAmount = _get(
-      this._record, 'details.assetCreate.initialPreissuedAmount'
+      this._record, 'details.createAsset.initialPreissuedAmount'
     )
 
     this.policies = this._policies()
     this.policy = this._policy()
 
-    this.details = _get(this._record, 'details.assetCreate.details')
-    this.terms = _get(this._record, 'details.assetCreate.details.terms')
-    this.termsKey = _get(this._record, 'details.assetCreate.details.terms.key')
+    this.details = _get(this._record, 'details.createAsset.details')
+    this.terms = _get(this._record, 'details.createAsset.details.terms')
+    this.termsKey = _get(this._record, 'details.createAsset.details.terms.key')
     this.termsName = _get(
-      this._record, 'details.assetCreate.details.terms.name'
+      this._record, 'details.createAsset.details.terms.name'
     )
     this.termsType = _get(
-      this._record, 'details.assetCreate.details.terms.type'
+      this._record, 'details.createAsset.details.terms.type'
     )
-    this.logo = _get(this._record, 'details.assetCreate.details.logo')
-    this.logoKey = _get(this._record, 'details.assetCreate.details.logo.key')
-    this.logoName = _get(this._record, 'details.assetCreate.details.logo.name')
-    this.logoType = _get(this._record, 'details.assetCreate.details.logo.type')
+    this.logo = _get(this._record, 'details.createAsset.details.logo')
+    this.logoKey = _get(this._record, 'details.createAsset.details.logo.key')
+    this.logoName = _get(this._record, 'details.createAsset.details.logo.name')
+    this.logoType = _get(this._record, 'details.createAsset.details.logo.type')
     this.externalSystemType = _get(
-      this._record, 'details.assetCreate.details.externalSystemType'
+      this._record, 'details.createAsset.details.externalSystemType'
     )
 
     this.attachedDetails = details
@@ -77,7 +80,7 @@ export class AssetCreateRequestRecord extends RequestRecord {
   }
 
   _policies () {
-    return _get(this._record, 'details.assetCreate.policies', [])
+    return _get(this._record, 'details.createAsset.policies', [])
       .map(policy => policy.value)
   }